Create grid.c
Create a grid that continually converges to a point below user-set value of epsilon.

Each point in the grid updates to the average of the four surrounding points. If the difference between the points in the old and new grids is below epsilon, print the grid and the number of iterations needed. Number of iterations is set by user but cannot exceed 100.